Move contracts back to their own "package"
Truffle is very rigid about its tests. It mandates that the truffle
tests are placed in the `test` directory, no exceptions, and it tries to
run all of the tests in that directory. In other words, we cannot put
any origin.js integration tests in that directory. We could work around
this, but at this point it starts getting messy, so I think it's
definitely best to just keep the truffle code separate for now.