Use config options rather than environment variables for ipfs settings
This is the proper way to set these options, because the platform will
never run as a standalone application. It will be imported into some
other environment. All options should be passed in explicitly into the
constructor.

This also makes local development easier, as environment variables can
be set in just one place - the demo dapp.