Improvements to README
* Start with step for Forking the repo, since that will be required for
all contributors without write access to the repo.

* Add correct URL for git cloning instructions (repo was renamed but
that name in the README did not change). This is duplicative of
https://github.com/OriginProtocol/origin-js/pull/117 but I am including
it because of the following bullet point.

* Suggest cloning with SSH instead of HTTPS since that is how most
people access GitHub repos, in my experience (could be wrong on this
one).

* Use `1` for all items in ordered list so that they do not need to be
renamed if items are added. GitHub markdown will automatically increment
numbers in ordered lists.