Rename contract Identity -> ClaimHolderPresigned
(1) Truffle tests for this contract aren't running for me with the name `identity`. I'm not exactly sure why but if the fix is changing the name, I'm just going to change the name.

(2) ClaimHolderPresigned is a more descriptive name. It's the same as ClaimHolder, except with the ability to accept presigned messages in the constructor. This name reflects how this contract relates to ClaimHolder, which can hopefully reduce confusion.