Refactor `user registry`/`claim holder` relationship
Now the user directly calls the identity contracts, and the identity contracts add themselves to the user registry.

This seems cleaner and simpler, but the real motivation to refactor was to allow identity contract addresses to be reliably predicted before they are actually deployed (for creating presigned claims). If the identity contract is deployed directly by the user, then its address can be predicted safely. If it is deployed by a shared user registry contract though, that opens the door to race conditions where other people might be deploying identities at the same time.