Only add new attestations in `set` method
This will allow consumers of this library to simply pass in the state that they want (both the profile and attestations). In other words, this efficiently replaces the existing user with whatever is passed in. (The only exception to this is removing attestations. We won't worry about that for now.)