Fix local vs. test net demo dapp config
It's been clarified already that we want the demo dapp to work out of the box on the test net, without any other local dependencies (local blockchain, local origin.js, local IPFS).

That said, we also want to make it as easy as possible to do local development.

So, with this config:
- Out of the box there are no local dependencies
- To get local decelopment config, just `cp .env.dev .env`