Go back to having just one purchase contract
This way we can have a single purchase contract that works across listings - any logic that depends on the type of listing should be delegated to the parent listing contract.

For example, isPaymentSufficient encapsulates the logic for whether a purchase has sufficient payment. The listing contract can simply compare against a price attribute, or it can be hard-coded to always return true (and defer to a separate approval process).

This makes purchase contracts agnostic of the type of listing they were spawned from. This allows our logic to be reused, provides a common interface, and hopefully makes it easier to work with in general.