Manage IPFS data in inherited rather than base contract
I thought this through a bit more, and realized that at some point we will probably also want to version other things besides IPFS hash (basically any other mutable data that we want to store on-chain). The data that we want to store in unit listings will probably be different than the data we want to store in fractional listings. So I'm moving the versioning stuff to the fractional listing contracts. We can always do something similar in the unit listing contracts - but I don't think we want them to be coupled. We want inherited contracts to be able to customize their own data structures.