Remove overloaded, unused event signature
We had an overloaded `ClaimAdded` event with a `signatureType` parameter. I'm not sure what the deal on this is - the parameter name is mentioned in some ERC735 comments (https://github.com/ethereum/EIPs/issues/735#issuecomment-335127715) but is not in the solidity interface or in mirceapasoi's implementation (https://github.com/mirceapasoi/erc725-735/blob/master/contracts/ERC735.sol). We are also not using the `signatureType` parameter.

I would have left this event there, but it is causing the `getPastEvents` lookup to fail when you search for `ClaimAdded`. No events are returned. Apparently, it gets confused by overloaded events and just doesn't return anything.

Simplest solution seems to be to remove the unused event signature.